pyproject.toml
[project]
name = "WebhookBot"
version = "0.1.0"
description = "WebhookBot"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "nonebot2[fastapi]>=2.5.0",
    "nonebot2[httpx]>=2.5.0",
    "nonebot2[websockets]>=2.5.0",
    "nonebot2[aiohttp]>=2.5.0",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-alconna>=0.61.1"
]

[tool.nonebot]
plugin_dirs = ["./plugins"]
builtin_plugins = []

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
"@local" = []

[tool.nonebot.plugins]
"@local" = []
nonebot-plugin-alconna = ["nonebot_plugin_alconna"]
